feat: create ciphertext storage on CREATE/CREATE2
Some storage backends (e.g. geth) require that the ciphertext storage
contract account is explicitly created. We do that only once on CREATE
and CREATE2 by initially setting its nonce from 0 to 1 and subsequently
not doing anything if nonce is not 0.

Above is not required for Ethermint, for example, but it doesn't impact
it either.